+++ internal/server/middleware/logger.go
@@ -7,39 +7,12 @@ import (
 	"time"
 )
 
-type LoggerOptions struct {
-	// SkipPaths defines which paths should not be logged
-	SkipPaths []string
-}
-
-// responseWriter captures the status code.
-type responseWriter struct {
-	http.ResponseWriter
-	status int
-}
-
-func newResponseWriter(w http.ResponseWriter) *responseWriter {
-	return &responseWriter{
-		ResponseWriter: w,
-		status:         http.StatusOK, // Default to 200 if WriteHeader is never called
-	}
-}
-
-func (rw *responseWriter) WriteHeader(statusCode int) {
-	rw.status = statusCode
-	rw.ResponseWriter.WriteHeader(statusCode)
-}
-
-// WithRequestLogging returns a middleware that logs request details using provided logger.
-func WithRequestLogging(logger *slog.Logger, opts *LoggerOptions) func(http.Handler) http.Handler {
-	if opts == nil {
-		opts = &LoggerOptions{}
-	}
-
+// RequestLogger returns a middleware that logs request details using provided logger.
+func RequestLogger(logger *slog.Logger, skipPaths []string) Middleware {
 	return func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			// Skip logging for specified paths
-			for _, path := range opts.SkipPaths {
+			for _, path := range skipPaths {
 				if strings.HasPrefix(r.URL.Path, path) {
 					next.ServeHTTP(w, r)
 					return
@@ -62,4 +35,22 @@ func WithRequestLogging(logger *slog.Logger, opts *LoggerOptions) func(http.Hand
 			)
 		})
 	}
+}
+
+// responseWriter captures the status code.
+type responseWriter struct {
+	http.ResponseWriter
+	status int
+}
+
+func newResponseWriter(w http.ResponseWriter) *responseWriter {
+	return &responseWriter{
+		ResponseWriter: w,
+	}
+}
+
+// WriteHeader captures the status code and calls the underlying ResponseWriter.
+func (rw *responseWriter) WriteHeader(statusCode int) {
+	rw.status = statusCode
+	rw.ResponseWriter.WriteHeader(statusCode)
 }
