+++ internal/server/middleware/rate_limiter.go
@@ -1,14 +1,25 @@
 package middleware
 
 import (
+	"fmt"
+	"log/slog"
 	"net"
 	"net/http"
+	"net/netip"
+	"strings"
 
 	"github.com/aerospike/aerospike-backup-service/v3/internal/util"
+	"github.com/aerospike/aerospike-backup-service/v3/pkg/model"
+	"golang.org/x/time/rate"
 )
 
-func RateLimiter(rateLimiter *util.IPRateLimiter, whiteList *util.IPWhiteList,
-) func(http.Handler) http.Handler {
+func RateLimiter(config *model.RateLimiterConfig) Middleware {
+	limiters := util.NewIPRateLimiter(
+		rate.Limit(config.GetTpsOrDefault()),
+		config.GetSizeOrDefault(),
+	)
+	whitelist := newIPWhiteList(config.GetWhiteListOrDefault())
+
 	return func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			ip, _, err := net.SplitHostPort(r.RemoteAddr)
@@ -17,15 +28,76 @@ func RateLimiter(rateLimiter *util.IPRateLimiter, whiteList *util.IPWhiteList,
 				return
 			}
 
-			if !whiteList.IsAllowed(ip) {
-				limiter := rateLimiter.GetLimiter(ip)
-				if !limiter.Allow() {
-					http.Error(w, http.StatusText(http.StatusTooManyRequests), http.StatusTooManyRequests)
-					return
-				}
+			if whitelist.isAllowed(ip) {
+				next.ServeHTTP(w, r)
+				return
+			}
+
+			limiter := limiters.GetLimiter(ip)
+			if !limiter.Allow() {
+				http.Error(w, http.StatusText(http.StatusTooManyRequests), http.StatusTooManyRequests)
+				return
 			}
 
 			next.ServeHTTP(w, r)
 		})
 	}
+}
+
+type IPWhiteList struct {
+	addresses map[string]*netip.Addr
+	networks  []*netip.Prefix
+	allowAny  bool
+}
+
+func newIPWhiteList(ipList []string) *IPWhiteList {
+	addresses := make(map[string]*netip.Addr)
+	networks := make([]*netip.Prefix, 0)
+	var allowAny bool
+
+	for _, ip := range ipList {
+		if strings.HasPrefix(ip, "0.0.0.0") {
+			allowAny = true
+			continue
+		}
+		network, err := netip.ParsePrefix(ip)
+		if err != nil {
+			ipAddr, err := netip.ParseAddr(ip)
+			if err != nil {
+				panic(fmt.Sprintf("invalid ip configuration: %s", ip))
+			}
+			addresses[ip] = &ipAddr
+		} else {
+			networks = append(networks, &network)
+		}
+	}
+
+	return &IPWhiteList{
+		addresses: addresses,
+		networks:  networks,
+		allowAny:  allowAny,
+	}
+}
+
+func (wl *IPWhiteList) isAllowed(ip string) bool {
+	if wl.allowAny {
+		return true
+	}
+	ipAddr, err := netip.ParseAddr(ip)
+	if err != nil {
+		slog.Warn("Invalid client ip", "ip", ip)
+		return false
+	}
+	_, ok := wl.addresses[ip]
+	if ok {
+		return true
+	}
+
+	for _, network := range wl.networks {
+		if network.Contains(ipAddr) {
+			return true
+		}
+	}
+
+	return false
 }
