+++ internal/server/server.go
@@ -9,9 +9,7 @@ import (
 
 	"github.com/aerospike/aerospike-backup-service/v3/internal/server/handlers"
 	"github.com/aerospike/aerospike-backup-service/v3/internal/server/middleware"
-	"github.com/aerospike/aerospike-backup-service/v3/internal/util"
 	"github.com/aerospike/aerospike-backup-service/v3/pkg/model"
-	"golang.org/x/time/rate"
 )
 
 const (
@@ -27,13 +25,6 @@ type HTTPServer struct {
 func NewHTTPServer(serverConfig *model.HTTPServerConfig, h *handlers.Service, logger *slog.Logger) *HTTPServer {
 	addr := fmt.Sprintf("%s:%d", serverConfig.GetAddressOrDefault(), serverConfig.GetPortOrDefault())
 
-	rateLimiterConfig := serverConfig.GetRateOrDefault()
-	rateLimiter := util.NewIPRateLimiter(
-		rate.Limit(rateLimiterConfig.GetTpsOrDefault()),
-		rateLimiterConfig.GetSizeOrDefault(),
-	)
-	whitelist := util.NewIPWhiteList(rateLimiterConfig.GetWhiteListOrDefault())
-
 	// Create router
 	mux := NewServeMux(
 		fmt.Sprintf("/%s", restAPIVersion),
@@ -41,14 +32,9 @@ func NewHTTPServer(serverConfig *model.HTTPServerConfig, h *handlers.Service, lo
 		h,
 	)
 
-	// Configure logging middleware
-	loggerOpts := &middleware.LoggerOptions{
-		SkipPaths: []string{"/health", "/ready", "/metrics"},
-	}
-
-	// Apply middleware chain
-	handler := middleware.WithRequestLogging(logger, loggerOpts)(
-		middleware.RateLimiter(rateLimiter, whitelist)(mux),
+	handler := middleware.Wrap(mux,
+		middleware.RateLimiter(serverConfig.GetRateOrDefault()),
+		middleware.RequestLogger(logger, []string{"/health", "/ready", "/metrics"}),
 	)
 
 	return &HTTPServer{
